requirements.txt
alembic==1.12.0
attrs==23.2.0
bidict==0.22.1
blinker==1.6.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
decorator==4.4.2
dnspython==2.4.2
dominate==2.9.1
email-validator==2.0.0.post2
Flask==2.3.3
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
gunicorn==22.0.0
h11==0.14.0
idna==3.4
imageio==2.34.1
imageio-ffmpeg==0.4.9
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
moviepy==1.0.3
mysql-connector==2.2.9
mysql-connector-python==8.4.0
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.0
pillow==10.3.0
proglog==0.1.10
psycopg2-binary
pycparser==2.22
PySocks==1.7.1
python-engineio==4.7.1
python-socketio==5.9.0
pytz==2024.1
requests==2.31.0
selenium==4.20.0
simple-websocket==0.10.1
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.20
tqdm==4.66.2
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==2.0.6
visitor==0.1.3
Werkzeug==2.3.7
wsproto==1.2.0
WTForms==3.0.1
